Rebase Blood Drinker System, Arachne, Oneirophage (#438)
This is a simple rebase of the Blood Drinker System, and its related
features that have been commented out and/or omitted due to its lack of
rebase. I am NOT substantially updating any of this code at this time,
outside of the barest minimum updates needed to make it run in the first
place. The reason I am doing this is that I require the Blood Drinker
system functional as a prerequisite for future features, and I will
update or refactor it when needed.

Arachne are still pending a Full Rework, but that is beyond the scope of
this PR.

- [x] Make the code functional
- [x] Port Arachne
- [x] Uncomment Oneirophages
- [x] Re-add Oneirophage midround event

:cl:
- add: Arachne have been reimplemented!
- add: Oneirophages are back!

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>